Digit counting now working

+++ include/nanofmt/serialize.h
@@ -7,7 +7,7 @@
 namespace nanofmt { namespace nanofmt_detail {
 
 
-void serialize(char* templateStr, unsigned arg, RepFieldData repFieldData);
+void serialize(char* templateStr, unsigned long long arg, RepFieldData repFieldData);
 // void serialize(char* templateStr, float arg, RepFieldData repFieldData);
 // void serialize(char* templateStr, const char* arg, RepFieldData repFieldData);
